requirements.txt
Flask==3.1.0
flask-cors==5.0.1
yt-dlp==2025.3.31
ffmpeg-python==0.2.0
openai-whisper==20240930
torch==2.6.0
transformers==4.51.3
sentencepiece==0.2.0
sentence-transformers==4.1.0
spacy==3.8.5
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
keybert==0.9.0
scikit-learn==1.6.1
numpy==2.2.4
tqdm==4.67.1
